Fixes wallet tests wrt dumpwallet RPC command
The tests assumed that a legacy wallet was created with the `createwallet` RPC
command, but since v23.0 the command creates native descriptor wallet which is
not compatible with the `dumpwallet` command. Since the `dumpwallet` command
only works for legacy wallets it makes sense to always create a legacy wallet
for in the tests for it.