block difficulty to use double instead of arbitrary precision
Back in #14 I added the difficulty value to the block header response as a
`Scientific` number for arbitrary precision, however I now think this was an
oversight and instead should have been a `double`. This is because in
bitcoin-core the block difficulty value corresponds to a `double` therefore it
doesn't make sense for us to provide this number as being of arbitrary
precision when in reality it isn't.