Relax servant-client constraints
It seems that back in #15 I set too tight the constraints on servant-client,
downstream consumers are not able to find a build plan when using the stackage
snapshot LTS 21.22 for the simple reason that `servant-client-0.19.1` does not
exist, there's however a version of `servant-0.19.1` that when pinning
`servant-client` to the released sources of that version will make everything
work just fine, but only if we have `>=0.19` and not `>=0.19.1` (because
servant-client itself doesn't have that version).

Downstream users on stackage LTS 21.22 need to pin servant-client at
58aa0d1c0c19f7b1c26ffc52bfd65c70934704c9, strictly speaking we do not need the
pin here because `servant-client-0.20` will be used and that works just fine,
it's only if the build plan selects `servant-client-0.19` that we need this.

I put a note about this in the README.